+++ relay-server/src/envelope.rs
@@ -116,6 +116,8 @@ pub enum ItemType {
     CheckIn,
     /// A standalone span.
     Span,
+    /// A standalone OpenTelemetry span.
+    OtelSpan,
     /// UserReport as an Event
     UserReportV2,
     /// A new item type that is yet unknown by this version of Relay.
@@ -165,6 +167,7 @@ impl fmt::Display for ItemType {
             Self::ReplayRecording => write!(f, "replay_recording"),
             Self::CheckIn => write!(f, "check_in"),
             Self::Span => write!(f, "span"),
+            Self::OtelSpan => write!(f, "otel_span"),
             Self::Unknown(s) => s.fmt(f),
         }
     }
@@ -196,6 +199,7 @@ impl std::str::FromStr for ItemType {
             "replay_recording" => Self::ReplayRecording,
             "check_in" => Self::CheckIn,
             "span" => Self::Span,
+            "otel_span" => Self::OtelSpan,
             other => Self::Unknown(other.to_owned()),
         })
     }
@@ -604,7 +608,8 @@ impl Item {
             ItemType::ClientReport => None,
             ItemType::CheckIn => Some(DataCategory::Monitor),
             ItemType::Unknown(_) => None,
-            ItemType::Span => None, // No outcomes, for now
+            ItemType::Span => None,     // No outcomes, for now
+            ItemType::OtelSpan => None, // No outcomes, for now
         }
     }
 
@@ -789,7 +794,8 @@ impl Item {
             | ItemType::ReplayRecording
             | ItemType::Profile
             | ItemType::CheckIn
-            | ItemType::Span => false,
+            | ItemType::Span
+            | ItemType::OtelSpan => false,
 
             // The unknown item type can observe any behavior, most likely there are going to be no
             // item types added that create events.
@@ -824,6 +830,7 @@ impl Item {
             ItemType::Profile => true,
             ItemType::CheckIn => false,
             ItemType::Span => false,
+            ItemType::OtelSpan => false,
 
             // Since this Relay cannot interpret the semantics of this item, it does not know
             // whether it requires an event or not. Depending on the strategy, this can cause two